     \name{gull}
     \docType{data}
     \alias{gull}
     \title{GPS data points from a lesser black-backed gull }
     \description{
       This data set contains a collection of GPS coordinates
	   obtained between June 2008 and July 2009.  The GPS tag,
	   developed by the University of Amsterdam, was worn by
	   a lesser black-backed gull (Larus fuscus).
     }
     \usage{gull}
     \format{A data.frame containing 22115 observations of the varibles latitude, longitude, altitude, and datetime.}
     \keyword{datasets}
